Tests seem to run but take forever
Adjust the configuration of `settings_tests.py` to use the MySQL
instance on the other docker container (similarly to how the dev setup
does it). Also copy the `tox` file(s) to the docker server container, so
that tox can be run from within it.

To run the tests, start the container (`docker compose up`),  enter an
interactive shell in the container (`docker exec -it <id of badgr
server> bash`) and run `tox` there.

Currently I'm not sure, if the tests *actually* run, because they don't
seem to complete ever.