Fix IntegrityError
The error is prevented by updating the relevant entity if it already
exists, instead of always trying to create a new one. This does not yet
solve unexpected behavior when changing the role of the issuer staff
that created the issuer: When changing it it persists that change, but
when saving again afterwards, it switches back to "Owner". It's probably
best to prohibit changes to the role of the creator of an issuer
altogether.