Add and configure mozilla-django-oidc
The views that are provided by this library are not yet used. We need to
expose them to the badgr-ui, make the relevant requests in the ui and
probably create an interface in the UI that allows for adding an E-Mail
address to the account that is authenticated via OIDC. For this also
further configurations to the library have to be made, because currently
it will try to automatically connect OIDC accounts to django users via
the E-Mail address.