Store session ID alongside access token
This adds a one to one table storing the session ID belonging to access
tokens created after OIDC authentication. This allows for a triggered
logout via OIDC, since the LOGOUT_TOKEN also contains this session ID.
The creation is tested, the logout can't be tested yet though, because
keycloak first has to be configured to also use our endpoint.