Check that only verified issuers create badges
This is the backend check corresponding to the frontend button
disabling. This check throws an error, in the end bubbling in a 500
error if the user somehow manages to invoke a badge creation with an
unverified issuer. This shows as a simple error message in the frontend
(if DEBUG is disabled).