feat: transfer ownership on user deletion
If a user is deleted the ownership of all associated issuers is
transferred to the next staff with the highest privileges. Also the
contact email of the issuer is set to the primary email of the new
owner. "Owner" in this sense also implicates creator, to ensure that no
issuer exists without a creator (even though this is kind of a lie).