feat: also adjust mail address if owner removed
So far the changes were only regarding the owner getting *deleted*, not
*removed*. With these changes the contact mail address is also adjusted
if the user with the current contact mail is removed from the
institution. This was basically already implemented, the delete method
just had to be called one by one, the creator / owner relationship had
to be adjusted a bit in the owner retaining method and some cache issues
had to be fixed.